+++ scripts/phenomic-theme-base.js
@@ -1,7 +1,7 @@
 import lnfs from "lnfs"
 import testFolder from "./utils/test-folder.js"
 
-const target = "phenomic-theme-base"
+const target = "themes/phenomic-theme-base"
 testFolder(
   target,
   {
@@ -9,7 +9,10 @@ testFolder(
 
     // phenomic-theme-base cleanup
     cleanup: async () => Promise.all([
-      await lnfs("docs/package.json", "phenomic-theme-base/package.json"),
+      await lnfs(
+        "docs/package.json",
+        "themes/phenomic-theme-base/package.json"
+      ),
     ]),
   }
 )