+++ themes/phenomic-theme-base/scripts/phenomic.browser.js
@@ -1,8 +1,8 @@
 import "whatwg-fetch"
 
-import metadata from "../web_modules/app/metadata"
-import routes from "../web_modules/app/routes"
-import store from "../web_modules/app/store"
+import metadata from "../src/metadata.js"
+import routes from "../src/routes.js"
+import store from "../src/store.js"
 
 import phenomicClient from "phenomic/lib/client"