+++ themes/phenomic-theme-base/scripts/phenomic.node.js
@@ -1,6 +1,6 @@
-import metadata from "../web_modules/app/metadata"
-import routes from "../web_modules/app/routes"
-import store from "../web_modules/app/store"
+import metadata from "../src/metadata.js"
+import routes from "../src/routes.js"
+import store from "../src/store.js"
 
 import phenomicStatic from "phenomic/lib/static"