+++ themes/phenomic-theme-base/src/layouts/Homepage/index.js
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from "react"
 import enhanceCollection from "phenomic/lib/enhance-collection"
 
 import Page from "../Page"
-import PagesList from "../../PagesList"
+import PagesList from "../../components/PagesList"
 
 const numberOfLatestPosts = 6