+++ themes/phenomic-theme-base/src/metadata.js
@@ -1,5 +1,5 @@
 // you can add any STATIC data you want here
-import pkg from "../../package.json"
+import pkg from "../package.json"
 
 export default {
   pkg,