+++ themes/phenomic-theme-base/src/routes.js
@@ -1,14 +1,14 @@
 import React, { Component } from "react"
 import { Route } from "react-router"
 
-import LayoutContainer from "../LayoutContainer"
-import PhenomicPageContainer from "phenomic/lib/PageContainer"
+import AppContainer from "./AppContainer"
+import { PageContainer as PhenomicPageContainer } from "phenomic"
 
-import Page from "../layouts/Page"
-import PageError from "../layouts/PageError"
-import PageLoading from "../layouts/PageLoading"
-import Homepage from "../layouts/Homepage"
-import Post from "../layouts/Post"
+import Page from "./layouts/Page"
+import PageError from "./layouts/PageError"
+import PageLoading from "./layouts/PageLoading"
+import Homepage from "./layouts/Homepage"
+import Post from "./layouts/Post"
 
 class PageContainer extends Component {
   render() {
@@ -29,7 +29,7 @@ class PageContainer extends Component {
 }
 
 export default (
-  <Route component={ LayoutContainer }>
+  <Route component={ AppContainer }>
     <Route path="*" component={ PageContainer } />
   </Route>
 )